backEnd/requirements.txt
flask
flask-cors
sqlalchemy
flask-sqlalchemy
gunicorn
python-dotenv
psycopg2-binary